Add .gitignore, package.json, and modules for Vercel Image Optimization
Added a .gitignore file to prevent node modules and package-lock.json file from being tracked. Introduced a package.json file for the "lit-vercel-index" project. This project is aimed at image optimization for Vercel with Lit, hence the creation of index.mjs and define.mjs files which define a LitElement for image loading, scaling, and optimization on Vercel platforms.