Add VercelImageGenerator to handle image URLs
The VercelImageGenerator class was created and added to the code base in generator.mjs to handle the generation of image URLs, moving the logic out of the `get url` method in index.mjs. This separates responsibilities between classes increasing modularity. I also updated package.json to include the new generator.mjs module in exports and files. A `connectedCallback` was added in index.mjs to instantiate the VercelImageGenerator with the current context. This represents an overall refactoring to improve the organization and cleanliness of the code.